voice-response-llm/requirements.txt
streamlit
ollama
openai-whisper
pyttsx3
ffmpeg-python